move ResourceState to its own package
