+++ bundle/statemgmt/resourcestate/resourcestate.go
@@ -1,4 +1,5 @@
-package statemgmt
+// This is in a separate package to avoid import cycles because it is imported by both terraform and statemgmt.
+package resourcestate
 
 type ResourceState struct {
 	ID string
